requirements.txt
numpy
python-binance
pandas
dash
plotly
matplotlib
backtrader
websocket-client
tabulate
tqdm